# api/openapi.yaml
openapi: 3.0.3
info:
  title: SCORM CMI API
  description: |
    A comprehensive REST API for managing SCORM 2004 courses and CMI (Computer Managed Instruction) data.
    
    This API provides endpoints for:
    - Uploading and managing SCORM packages
    - Creating and managing learner registrations
    - Tracking CMI data (progress, scores, interactions)
    - Generating learning analytics reports
    
    The API follows SCORM 2004 4th Edition specifications and supports all major CMI data elements.
  version: 1.0.0
  contact:
    name: SCORM CMI API Support
    email: support@example.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080/api
    description: Development server
  - url: https://your-domain.com/api
    description: Production server
paths:
...
  /registrations:
    get:
      summary: List registrations
      description: Retrieve a list of learner registrations, optionally filtered by learner
      tags:
        - Registrations
      parameters:
        - name: learner_id
          in: query
          required: false
          schema:
            type: string
          description: Filter registrations by learner ID
      responses:
        '200':
          description: List of registrations retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    registration:
                      $ref: '#/components/schemas/Registration'
                    course_title:
                      type: string
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      summary: Create a new registration
      description: Enroll a learner in a course
      tags:
        - Registrations
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationCreateRequest'
      responses:
        '201':
          description: Registration created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Course not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Registration already exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /registrations/{registrationId}:
    get:
      summary: Get registration details
      description: Retrieve detailed information about a specific registration
      tags:
        - Registrations
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier of the registration
      responses:
        '200':
          description: Registration details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '404':
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      summary: Update registration
      description: Update registration fields such as status, score, etc.
      tags:
        - Registrations
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier of the registration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegistrationUpdateRequest'
      responses:
        '200':
          description: Registration updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Registration'
        '400':
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      summary: Delete registration
      description: Remove a registration from the system
      tags:
        - Registrations
      parameters:
        - name: registrationId
          in: path
          required: true
          schema:
            type: string
          description: Unique identifier of the registration
      responses:
        '204':
          description: Registration deleted successfully
        '404':
          description: Registration not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  schemas:
...
    Registration:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier
        course_id:
          type: string
          description: Associated course ID
        learner_id:
          type: string
          description: Learner's unique identifier
        learner_name:
          type: string
          description: Learner's display name
        status:
          type: string
          enum: ["not_attempted", "incomplete", "completed", "failed", "passed", "browsed"]
          description: Current lesson status
        score:
          type: number
          nullable: true
          description: Raw score
        min_score:
          type: number
          nullable: true
          description: Minimum passing score
        max_score:
          type: number
          nullable: true
          description: Maximum possible score
        session_time:
          type: string
          description: Time spent in current session
        total_time:
          type: string
          description: Total time spent on course
        location:
          type: string
          description: Bookmark/location in course
        suspend_data:
          type: string
          description: Suspended session data
        exit:
          type: string
          description: Exit condition
        entry:
          type: string
          description: Entry condition
        created_at:
          type: string
          format: date-time
          description: Registration creation timestamp
        updated_at:
          type: string
          format: date-time
          description: Last update timestamp
        completed_at:
          type: string
          format: date-time
          nullable: true
          description: Completion timestamp
      required:
        - id
        - course_id
        - learner_id
        - learner_name
        - status
        - created_at
        - updated_at
    RegistrationCreateRequest:
      type: object
      properties:
        course_id:
          type: string
          description: Course to enroll in
        learner_id:
          type: string
          description: Learner's unique identifier
        learner_name:
          type: string
          description: Learner's display name
      required:
        - course_id
        - learner_id
        - learner_name
    RegistrationUpdateRequest:
      type: object
      properties:
        status:
          type: string
          enum: ["not_attempted", "incomplete", "completed", "failed", "passed", "browsed"]
          description: Lesson status
        score:
          type: number
          description: Raw score
        min_score:
          type: number
          description: Minimum passing score
        max_score:
          type: number
          description: Maximum possible score
        session_time:
          type: string
          description: Session time
        total_time:
          type: string
          description: Total time
        location:
          type: string
          description: Bookmark location
        suspend_data:
          type: string
          description: Suspend data
        exit:
          type: string
          description: Exit condition
        entry:
          type: string
          description: Entry condition
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: Error message
      required:
        - error
...